requirements.txt
#requirements.txt
#langchain>=0.2.0
#langchain-community>=0.2.0
#langchain-openai>=0.1.7
#psycopg2-binary>=2.9.9
#SQLAlchemy>=2.0.30
#python-dotenv>=1.0.1
#pandas>=2.2.2
#tabulate>=0.9.0
#langgraph>=0.0.70



# Core Agent + LangGraph
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7
langgraph>=0.1.0
langgraph-checkpoint==1.0.12

# Database & ORM
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.30

# Data handling
pandas>=2.2.2
tabulate>=0.9.0

# Environment
python-dotenv>=1.0.1

# API Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
